requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.2.0=pypi_0
aiohttp=3.8.1=pypi_0
aiosignal=1.2.0=pypi_0
async-timeout=4.0.2=pypi_0
asynctest=0.13.0=pypi_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2022.6.15=py37h06a4308_0
charset-normalizer=2.1.0=pypi_0
click=8.1.3=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cycler=0.11.0=pypi_0
datasets=2.4.0=pypi_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
dill=0.3.5.1=pypi_0
entrypoints=0.4=py37h06a4308_0
evaluate=0.2.2=pypi_0
filelock=3.8.0=pypi_0
fonttools=4.34.4=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.1=pypi_0
fsspec=2022.7.1=pypi_0
future=0.18.2=pypi_0
giflib=5.2.1=h7b6447c_0
google-auth=2.10.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.47.0=pypi_0
huggingface-hub=0.8.1=pypi_0
idna=3.3=pypi_0
importlib-metadata=4.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py37h06a4308_0
ipython=7.31.1=py37h06a4308_1
jedi=0.18.1=py37h06a4308_1
jieba=0.42.1=pypi_0
joblib=1.1.0=pypi_0
jpeg=9b=h024ee3a_2
jupyter_client=7.2.2=py37h06a4308_0
jupyter_core=4.10.0=py37h06a4308_0
kiwisolver=1.4.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.1.0=h2733197_1
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
loguru=0.6.0=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.1=py37hd3c417c_0
mkl_random=1.2.2=py37h51133e4_0
multidict=6.0.2=pypi_0
multiprocess=0.70.13=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py37h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nltk=3.7=pypi_0
numpy=1.21.5=py37h6c91a56_3
numpy-base=1.21.5=py37ha15fc14_3
oauthlib=3.2.0=pypi_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.3.5=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py37hace64e9_1
pip=22.1.2=py37h06a4308_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.4=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pyarrow=9.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydeprecate=0.3.1=pypi_0
pygments=2.11.2=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
python=3.7.13=h12debd9_0
python-dateutil=2.8.2=pyhd3eb1b0_0
pytorch=1.7.1=py3.7_cuda10.1.243_cudnn7.6.3_0
pytorch-lightning=1.5.10=pypi_0
pytz=2022.2=pypi_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py37h6a678d5_0
readline=8.1.2=h7f8727e_1
regex=2022.7.25=pypi_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rouge=1.0.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.11.2=pypi_0
sentencepiece=0.1.97=pypi_0
setuptools=59.5.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.2=h5082296_0
tabulate=0.8.10=pypi_0
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.12.1=pypi_0
torchaudio=0.7.2=py37
torchmetrics=0.9.3=pypi_0
torchvision=0.8.2=py37_cu101
tornado=6.1=py37h27cfd23_0
tqdm=4.64.0=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.21.1=pypi_0
typing_extensions=4.3.0=py37h06a4308_0
urllib3=1.26.11=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xxhash=3.0.0=pypi_0
xz=5.2.5=h7f8727e_1
yarl=1.8.1=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.8.1=pypi_0
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=haebb681_0
